requirements.txt
Django>=2.0
psycopg2
djangorestframework==3.6.4
